+++ apps/site-landing/src/components/Prismic/index.tsx
@@ -1,7 +1,2 @@
-
-export const Help = () => {
-  return null;
-}
-
 export * from './reducer'
 export * from './types'