+++ apps/site-nft/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.45](https://github.com/thecointech/thecoin/compare/v0.2.43...v0.2.45) (2021-10-15)
+
+**Note:** Version bump only for package @thecointech/site-nft
+
+
+
+
+
+## [0.2.44](https://github.com/thecointech/thecoin/compare/v0.2.43...v0.2.44) (2021-10-15)
+
+**Note:** Version bump only for package @thecointech/site-nft
+
+
+
+
+
+## [0.2.43](https://github.com/thecointech/thecoin/compare/v0.2.42...v0.2.43) (2021-10-14)
+
+**Note:** Version bump only for package @thecointech/site-nft
+
+
+
+
+
 ## [0.2.42](https://github.com/thecointech/thecoin/compare/v0.2.41...v0.2.42) (2021-10-12)
 
 **Note:** Version bump only for package @thecointech/site-nft